requirements_demo.txt
# 智能辅助诊断平台演示系统依赖

# Gradio Web框架
gradio>=4.0.0

# 可视化
plotly>=5.0.0
scikit-image>=0.19.0

# 医学图像处理
SimpleITK>=2.0.0
h5py>=3.0.0
nibabel>=4.0.0

# 数值计算 (项目已有)
# numpy>=1.21.0
# scipy>=1.7.0
# torch>=1.12.0

# 其他
PyYAML>=6.0
tqdm>=4.60.0
